Add files via upload
Assume instead of x as a double you input three integers hours, minutes and seconds that the user talked. For example input “2 5 45” means 02:05.45, 2 hours 5 minutes 45 seconds. Your multipliers and decisions remains the same, however smallest units that company bills is 8 seconds. That means you should round the talk amount to closest greater value divisible by 8. For example if user spoke 1 minute 2 seconds, it is 62 seconds and you should round it up to 64, first value divisible to 8. Then it is 1 minute 4 seconds, and make your calculations based on that. (note that a minute is 60 seconds, you should make TL per minute calculation based on that).